v1.0: Use full serial number for monitor identification
The serial number as viewed in the UI is a 7 digit number. The packet format splits that over two fields — the first 2 digits go in “device ID”, and the last 5 in “serial number”. Previously we were only using the packet “serial number” field to identify a monitor, but this is technically incorrect.

This is a breaking change, so we bump the major version.