Align watt-second/kilowatt-hour calculations with watts
We calculate watts as delta(consumed) - delta(produced), where
delta(produced) could be 0 for non-net-metered channels, just based on
the data in the packet. But Ws and kWh were being computed differently,
using the net_metering setting on the channel to switch between them.
This difference is unnecessary and confusing, so eliminating it.